Follows HTTP 301, 302, 303, 307, and 308 redirects.
The implementation prior to this commit doesn't take into account when a
repository is moved in Github and will return an unssuccesful response
with appropriate headers and content to indicate that the resource was moved.

This commits adds a Faraday middleware taken from lostisland/faraday_middleware
and inserts it in the middleware pipeline after the request middlewares and
before all the current reponse middlewares.